# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The complete API documentation of the DentalHQ API",
        "version": "1.0.0",
        "title": "DentalHQ",
        "termsOfService": "```\n@Copyright 2017 DentalHQ\n```",
        "contact": {}
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json",
        "application/pdf"
    ],
    "produces": [
        "application/json",
        "application/pdf"
    ],
    "paths": {
...
        "/api/v1/dentist/{officeId}/{year}/{month}/general": {
            "get": {
                "tags": [
                    "Admin Statistics"
                ],
                "summary": "General Office Revenue Report",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "The JWT",
                        "x-example": "JWT xxxxx.yyyyy.zzzzz"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The PDF report"
                    }
                }
            },
            "parameters": [
                {
                    "name": "officeId",
                    "in": "path",
                    "required": true,
                    "type": "number",
                    "description": "The ID of the office to draw the report on"
                },
                {
                    "name": "year",
                    "in": "path",
                    "required": true,
                    "type": "number",
                    "description": "The year for which the report is compiled from",
                    "x-example": 2017
                },
                {
                    "name": "month",
                    "in": "path",
                    "required": true,
                    "type": "number",
                    "description": "The month for which the report is compiled from (1 - 12)",
                    "x-example": 2
                }
            ]
        },
        "/api/v1/reports/dentists/{year}/{month}": {
            "get": {
                "tags": [
                    "Admin Statistics"
                ],
                "summary": "Retrieve Master Report",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "The JWT",
                        "x-example": "JWT xxxxx.yyyyy.zzzzz"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The PDF report"
                    },
                    "403": {
                        "description": "Restricted to Admins only"
                    }
                }
            },
            "parameters": [
                {
                    "name": "year",
                    "in": "path",
                    "required": true,
                    "type": "number",
                    "description": "The year in which the report is to be labeled as"
                },
                {
                    "name": "month",
                    "in": "path",
                    "required": true,
                    "type": "number",
                    "description": "The month in which the report is to labeled as"
                }
            ]
        },
...
    },
...
}